Assign dragover class only when dragging over target element.
Previously commented out code did not remove the dragover class
from the target element when the drag left the area. This was because
an enter and leave events are fired when crossing sub element
boundaries, and the later leave event was leaving the class off
the element. Keeping a count allows the element to keep the class
until the final leave.